move seal function to new file
